Aug 21th 1st commit

+++ Cart-Service/src/main/java/com/manura/foodapp/CartService/repo/CartRepo.java
@@ -13,6 +13,6 @@ import com.manura.foodapp.CartService.Table.CartTable;
  *
  */
 @Repository
-public interface CartRepo extends ReactiveCrudRepository<CartTable, Long> {
+public interface CartRepo extends ReactiveCrudRepository<CartTable, Integer> {
 
 }
